+++ py/desitarget/mtl.py
@@ -15,7 +15,7 @@ import sys
 from astropy.table import Table
 from astropy.io import ascii
 import fitsio
-from time import time
+from time import time, sleep
 from datetime import datetime, timezone
 from glob import glob, iglob
 
@@ -49,7 +49,7 @@ mtldatamodel = np.array([], dtype=[
 
 # ADM columns to add to the mtl/zcat data models for the Main Survey.
 msaddcols = np.array([], dtype=[
-    ('Z_QN', '>f8'), ('DELTACHI2', '>f8'),
+    ('Z_QN', '>f8'), ('IS_QSO_QN', '>i2'), ('DELTACHI2', '>f8'),
     ])
 
 zcatdatamodel = np.array([], dtype=[
@@ -482,7 +482,7 @@ def make_mtl(targets, obscon, zcat=None, scnd=None,
     # NUMOBS_MORE should also be 0.
     # ## mtl['NUMOBS_MORE'] = ztargets['NUMOBS_MORE']
     ii = (priority <= 2)
-    log.info('{:d} of {:d} targets have priority zero, setting N_obs=0.'.format(
+    log.info('{:d} of {:d} targets have priority <=2, setting N_obs=0.'.format(
         np.sum(ii), n))
     ztargets['NUMOBS_MORE'][ii] = 0
 
@@ -1093,6 +1093,84 @@ def tiles_to_be_processed(zcatdir, mtltilefn, obscon, survey):
     return newtiles
 
 
+def make_zcat(zcatdir, tiles, obscon, survey):
+    """Make a catalog of redshifts used to inform the MTL loop.
+
+    Parameters
+    ----------
+    zcatdir : :class:`str`
+        Full path to the "daily" directory that hosts redshift catalogs.
+    tiles : :class:`~numpy.array`
+        Numpy array of tiles to be processed. Must contain at least:
+        * TILEID - unique tile identifier.
+        * ZDATE - final night processed to complete the tile (YYYYMMDD).
+    obscon : :class:`str`
+        A string matching ONE obscondition in the desitarget bitmask yaml
+        file (i.e. `desitarget.targetmask.obsconditions`), e.g. "DARK".
+        Governs how ZWARN is updated using `DELTACHI2` when `survey` is
+        "sv3" (in :func:`~desitarget.mtl.make_zcat_rr_backstop()`).
+    survey : :class:`str`, optional, defaults to "main"
+        Used to update `ZWARN` using `DELTACHI2` for a given survey type.
+        Options are ``'main'`` and ``'svX``' (where X is 1, 2, 3 etc.)
+        for the main survey and different iterations of SV, respectively.
+
+    Returns
+    -------
+    :class:`~astropy.table.Table`
+        A zcat in the official format (`zcatdatamodel`) compiled from
+        the `tiles` in `zcatdir`.
+
+    Notes
+    -----
+    - For surveys prior to "main" this is just a wrapper on
+      :func:`~desitarget.mtl.make_zcat_rr_backstop()`.
+    """
+    if survey != "main":
+        return make_zcat_rr_backstop(zcatdir, tiles, obscon, survey)
+
+    # ADM the root directory in the data model.
+    rootdir = os.path.join(zcatdir, "tiles", "cumulative")
+
+    # ADM for each tile, read in the spectroscopic and targeting info.
+    allzs = []
+    for tile in tiles:
+        # ADM build the correct directory structure.
+        tiledir = os.path.join(rootdir, str(tile["TILEID"]))
+        ymdir = os.path.join(tiledir, tile["ZDATE"])
+        # ADM and retrieve the zqso/lyazcat catalog.
+        qsozcatfns = sorted(glob(os.path.join(ymdir, "zqso*")))
+        for qsozcatfn in qsozcatfns:
+            zz = fitsio.read(qsozcatfn, "QSOZCAT")
+            allzs.append(zz)
+            # ADM check the correct TILEID was written in the fibermap.
+            if set(zz["ZTILEID"]) != set([tile["TILEID"]]):
+                msg = "Directory and fibermap don't match for tile".format(tile)
+                log.critical(msg)
+                raise ValueError(msg)
+    zs = np.concatenate(allzs)
+
+    # ADM remove -ve TARGETIDs which should correspond to sky fibers.
+    zs = zs[zs["TARGETID"] >= 0]
+
+    # ADM check the TARGETIDs are unique. If they aren't the likely
+    # ADM explanation is that overlapping tiles (which could include
+    # ADM duplicate targets) are being processed.
+    if len(zs) != len(set(zs["TARGETID"])):
+        msg = "a target is duplicated!!! You are likely trying to process "
+        msg += "overlapping tiles when one of these tiles should already have "
+        msg += "been processed and locked in mtl-done-tiles.ecsv"
+        log.critical(msg)
+        raise ValueError(msg)
+
+    # ADM write out the zcat as a file with the correct data model.
+    dm = survey_data_model(zcatdatamodel, survey=survey)
+    qsozcat = Table(np.zeros(len(zs), dtype=dm.dtype))
+    for col in qsozcat.dtype.names:
+        qsozcat[col] = zs[col]
+
+    return qsozcat
+
+
 def make_zcat_rr_backstop(zcatdir, tiles, obscon, survey):
     """Make a simple zcat using only redrock outputs.
 
@@ -1276,9 +1354,8 @@ def loop_ledger(obscon, survey='main', zcatdir=None, mtldir=None,
     if len(tiles) == 0:
         return hpdirname, mtltilefn, ztilefn, tiles
 
-    # ADM create the zcat: This will likely change, but for now let's
-    # ADM just use redrock.
-    zcat = make_zcat_rr_backstop(zcatdir, tiles, obscon, survey)
+    # ADM create the catalog of updated redshifts.
+    zcat = make_zcat(zcatdir, tiles, obscon, survey)
 
     # ADM insist that for an MTL loop with real observations, the zcat
     # ADM must conform to the data model. In particular, it must include
@@ -1294,12 +1371,19 @@ def loop_ledger(obscon, survey='main', zcatdir=None, mtldir=None,
     # ADM useful to know how many targets were updated.
     _, _, _, _, sky, _ = decode_targetid(zcat["TARGETID"])
     ntargs, nsky = np.sum(sky == 0), np.sum(sky)
-    log.info("Update state for {} targets (zcat also contains {} skies)".format(
-        ntargs, nsky))
+    msg = "Update state for {} targets".format(ntargs)
+    msg += " (the zcats also contain {} skies with +ve TARGETIDs)".format(nsky)
+    log.info(msg)
 
     # ADM update the appropriate ledger.
     update_ledger(hpdirname, zcat, obscon=obscon,
                   numobs_from_ledger=numobs_from_ledger)
+
+    # ADM for the main survey "holding pen" method, ensure the TIMESTAMP
+    # ADM in the mtl-done-tiles file is always later than in the ledgers.
+    if survey == "main":
+        sleep(1)
+        tiles["TIMESTAMP"] = get_utc_date(survey=survey)
 
     # ADM write the processed tiles to the MTL tile file.
     io.write_mtl_tile_file(mtltilefn, tiles)